fixup! refactor(core): Remove `LContainerFlags.HasChildViewsToRefresh`
